and, or: make them variadic
`And` and `Or` in Arel are really different: the former is variadic,
not the latter.  This was coped with by introducing a
`ArelExtensions::Nodes::Or` class, which derived from Arel::Nodes::And
to steal its variadic implementation.

That seems really clumsy.  Let's instead add variadic support to
`Arel::Nodes::Or`.

Denested children of the same kind, so that `And(And(a, b), And(c, d))`
is actually equal to `And(a, b, c, d)`.