mssql: date-diff: fix the way resolution is calculated
The assumption was: convert everything to seconds and let the db do the
addition.

This is problematic for large number, e.g. 10 years translate to
315569520 seconds, and apparently DBs don't like to add seconds this
way. It's better to specify it in years instead of seconds.

This is a fix for SQLServer, but I noticed the same behaviour in MySQL.