appveyor: restrict to ruby 2.5
Appveyor has multiple windows images. 2 of them are important for us:

- VS2015:
  - Has all SQL Server versions installed (2008 -> 2019)
  - Supports ruby up to 2.5
- VS2019
  - Has modern SQL Server versions installed (2017 -> 2019)
  - supports ruby 2.5+

I was unable to install ruby > 2.5 on VS2015, and I was unable to get
SQL Server 2008 -> 2012 installed on VS2019.

Since we already test for SQL Server 2017 -> 2019 on github's CI, we can
consider that testing with ruby 2.5 on the old versions of SQL server is
good enough.