test: fix the date test
Amusingly, the test passed at the time by mistake.

`created_at: nil` in the month of May will actually be populated
by activerecord with the current timestamp => dd-05-yyy …

When counting the # of users created in May, we get 10. However,
when the test runs in a different month, the number goes down
to 9, which only normal, but we missed it.

This should be a reminder that we need to run our CI daily, maybe even
twice or more if needed, especially that we have a lot of test depending
on date/time.