column_of: print the backtrace when schema reflection fails
Activerecord changes its internal implementaion of the schema cache of
the connection pool quite often, and it's understandable since they're
refactoring to separate concerns, and lately, to properly handle
concurrency in the pools (activerecord 7.1).

When discovering+fixing the issues, the backtrace was not shown, and the
error message we used to produce wasn't helpful, and worst, it was
unreadable.