dev: compose: use mysql 8
I was bitten by a very weird bug when trying to reproduce a failure in
the test suite for jdbc+mysql.

My docker env used mariadb, but the CI used mysql.

The CI connected to the db and failed at one particular assertion.

Locally, I couldn't even connect. After investigation, it turned out
that the tx_isolation is issued by default in the jdbc drivers.

So we better stick to the CI.

The reason why it was on mariadb was lost in translation between me and
other people, and it doesn't really matter I guess.

I have better ideas for the test suite, and they need to be addressed
once and for all.

If anything this particular experience shows not only how slow and
wastefult our CI is, and how unreliable our CI and dev environments are,
but also how blind it is for different database versions.